// generated/product/CouponUpdateResponse.ts
// Original file: src/product.proto


export interface CouponUpdateResponse {
  'type'?: (string);
  'couponCode'?: (string);
  'businessId'?: (string);
  'status'?: (string);
  'expiryDate'?: (string);
  'totalCoupons'?: (number);
  'remainingCoupons'?: (number);
  'businessName'?: (string);
  'message'?: (string);
}

export interface CouponUpdateResponse__Output {
  'type': (string);
  'couponCode': (string);
  'businessId': (string);
  'status': (string);
  'expiryDate': (string);
  'totalCoupons': (number);
  'remainingCoupons': (number);
  'businessName': (string);
  'message': (string);
}